# openapi.yml
...
openapi: 3.0.3
info:
  title: Uplink Connect API
  description: |
    Cloudflare-native data ingestion platform API.
    
    ## Authentication
    - **External endpoints**: Bearer token in `Authorization` header
    - **Internal endpoints**: `x-uplink-internal-key` header
  version: 3.01.0
  contact:
    name: BoringWorks
    url: https://github.com/Boring-Works/uplink-connect
servers:
  - url: https://uplink-core.codyboring.workers.dev
    description: Production Core API
  - url: https://uplink-edge.codyboring.workers.dev
    description: Production Edge API
...
paths:
...
  /internal/search/entities:
    post:
      tags: [Entities]
      summary: Search entities by similarity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                topK: { type: integer, default: 10 }
                filter: { type: object }
      responses:
        200:
          description: Search results
        503:
          description: Vector search not configured
  /internal/entities/{entityId}/lineage:
    get:
      tags: [Entities]
      summary: Get entity lineage
      parameters:
        - name: entityId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Entity lineage
        404:
          description: Entity not found

  # Artifacts
...
  /internal/dashboard/v2:
    get:
      tags: [Dashboard]
      summary: Dashboard data (v2)
      parameters:
        - name: window
          in: query
          schema: { type: integer, default: 86400 }
      responses:
        200:
          description: Dashboard data

  # Health Monitor
...
  /internal/settings:
    get:
      tags: [Settings]
      summary: Get platform settings
      responses:
        200:
          description: Platform settings
    put:
      tags: [Settings]
      summary: Update platform settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        200:
          description: Settings updated
  /internal/audit-log:
    get:
      tags: [Settings]
      summary: Get audit log
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: resourceType
          in: query
          schema: { type: string }
        - name: actor
          in: query
          schema: { type: string }
        - name: fromDate
          in: query
          schema: { type: string }
        - name: toDate
          in: query
          schema: { type: string }
      responses:
        200:
          description: Audit log entries

  # Browser
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    InternalKey:
      type: apiKey
      in: header
      name: x-uplink-internal-key
security:
  - InternalKey: []